Use double underscores for underlines
This changes underscore syntax to require two underscores (`__`) to
wrap text instead of just one underscore (`_`).

One of the benefits of this is that underscores in words, like a URL or
some type of slug-like string, will not be treated as special syntax,
being removed and converted to underlines. Underscores are pretty common
in Steam URLs, so this can be especially annoying.

If Steam allows double underscores in its URLs, and its fairly common,
this might need to be revisited.

This also addresses #247.

Closes #247